requirements.txt
# Core
pytz==2018.3
Django==2.0.3
django-configurations==2.0
gunicorn==19.7.1
newrelic==3.0.0.89

# For the persistence stores
# psycopg2==2.7.4
dj-database-url==0.5.0
psycopg2-binary==2.7.4


# Model Tools
django-model-utils==3.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.7.7
Markdown==2.6.11
django-filter==1.1.0

# Developer Tools
ipdb==0.11
ipython==6.3.0
mkdocs==0.17.3
flake8==3.5.0

# Testing
mock==2.0.0
factory-boy==2.10.0
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1

# Static and Media Storage
django-storages==1.6.6
boto3==1.6.22
django-extensions==2.0.0